requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==2.5.3
djangorestframework==3.9.2
get==2019.4.13
gunicorn==19.9.0
idna==2.8
mysql==0.0.2
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.0.0
post==2019.4.13
psycopg2-binary==2.8.2
public==2019.4.13
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
urllib3==1.24.2
